Made a bunch of critical fixes. Got the Itarget loss term working with the coil symmetries (using the QA example). Debugged divJ and found crucial error where Phi basis function 3 was incorrect ([x, 0, 0] instead of [z, 0, 0] so it had finite divergence). Fixed a whole host of errors with the flux jump constraints, and implemented the strategy for implementing the constraints suggest by Matt. Still occassionally having some issues getting a nonsingular constraint matrix, and next step is eventually to get some tests working that make sure the constraints are actually getting the flux jumps to match as desired.

+++ src/simsoptpp/winding_volume.h
@@ -12,4 +12,4 @@ using std::vector;
 Array winding_volume_geo_factors(Array& points, Array& integration_points, Array& plasma_normal, Array& Phi);
 std::tuple<Array, Array> winding_volume_flux_jumps(Array& coil_points, Array& Phi, double dx, double dy, double dz);
 Array make_winding_volume_grid(Array& normal_inner, Array& normal_outer, Array& dipole_grid_xyz, Array& xyz_inner, Array& xyz_outer);
-Array connections(Array& dipole_grid_xyz, int Nadjacent, int dx, int dy, int dz);
+Array connections(Array& dipole_grid_xyz, int Nadjacent, double dx, double dy, double dz);
